backend/requirements.txt
Flask==2.1.1
flask-cors==3.0.10
psycopg2-binary==2.9.3
Werkzeug==2.2.2